Added age-based immunities and applied them to the methods.
Now, every person has one of the 4 types of immunities, based on their age. Age of a person is generated with real-life age percentage. Now, every person gets sick based on their immunity number.

+++ Proje/core/src/com/mygdx/viralepidemicsim/FirstVersion/Person.java
@@ -63,7 +63,7 @@ abstract class Person extends SpriteBatch implements PersonInterfaces {
         return this.age;
     }
 
-    public int getImmunity() {
+    public byte getImmunity() {
         return this.immunity;
     }
